+++ tests/phpMyFAQ/ApiTest.php
@@ -13,7 +13,6 @@ use PHPUnit\Framework\TestCase;
  */
 class ApiTest extends TestCase
 {
-
     /** @var Configuration */
     private Configuration $configuration;
 
@@ -29,6 +28,7 @@ class ApiTest extends TestCase
 
     /**
      * @testdox return the available versions
+     * @throws Core\Exception
      */
     public function testGetVersions(): void
     {
@@ -52,6 +52,7 @@ class ApiTest extends TestCase
 
     /**
      * @testdox return the current verification hashes
+     * @throws Core\Exception
      */
     public function testGetVerificationIssues(): void
     {